Add first version of raising to llvm
This patch adds a converter that converts ptx to llvm.
For now, it just created an empty module with data layout
and target triple attributes. A unit test is also added to
verify this.

The code has also been refactored with the creating of an
include dir for the headers.

Finally, compile_to_ptx now has the ability to emit .ll files
and has more descriptive flags.