Add type inference to determine types of values
This patch adds an algorithm to determine the
type of the arguments of the original llvm function. The
idea is to traverse the graph in reverse propagating type
annotations all the way to the kernel arguments.

The test has been expanded to include most of the elementwise
add kernel. Also macros have been defined to simplify PTX definition
in the test.