This patch adds the ability to emit store, add, mul and alloca instructions
When generating store instructions, we first
check whether the location where we are storing to exists.
If not an alloca is generated.

We also emit mul and add instructions when the source
and destination operands are 32 bits long.

To enable using the kernel args, we do analysis to propagate
the kernel arg values through the address space instructions
(cvta) which are not relevant for LLVM IR.

Finally, we refactor the code and move common functions to utils.